+++ mediawiki/mediawiki.py
@@ -642,7 +642,6 @@ class MediaWiki(object):
             if results is not None and results - returned_results < max_pull:
                 search_params['cmlimit'] = results - returned_results
 
-            # print(last_cont)
         # end while loop
 
         if subcategories:
@@ -707,11 +706,8 @@ class MediaWiki(object):
             else:
                 parent_cats = categories[cat].categories
 
-            for pcat in parent_cats:
-                tree[cat]['parent-categories'].append(pcat)
-
-            for link in links[cat][0]:
-                tree[cat]['links'].append(link)
+            tree[cat]['parent-categories'].extend(parent_cats)
+            tree[cat]['links'].extend(links[cat][0])
 
             if depth and level >= depth:
                 for ctg in links[cat][1]:
@@ -865,7 +861,7 @@ class MediaWiki(object):
         ''' check for default error messages and throw correct exception '''
         if 'error' in response:
             http_error = ['HTTP request timed out.', 'Pool queue is full']
-            geo_error = ['Page coordinates unknown',
+            geo_error = ['Page coordinates unknown.',
                          ('One of the parameters gscoord, gspage, gsbbox is '
                           'required'), 'Invalid coordinate provided']
             err = response['error']['info']